refactor: streamline DiagramCanvas and FlowToolbar components by consolidating state management and removing unused imports
